Trivial refactoring: reshuffle types in files
